rename error-handling hook
